requirements.txt
dlib == 19.24.0
face_recognition == 1.3.0
Flask == 2.2.2
opencv-python == 4.6.0.66
py-postgresql == 1.2.2
